Reorganize type files
